requirements.txt
# Core
Django>=5.1,<6.0
django-environ>=0.11.2

# Auth & Security
bleach>=6.1

# Database
dj-database-url>=2.1
psycopg2-binary>=2.9

# REST API
djangorestframework>=3.15
django-filter>=24.1
django-cors-headers>=4.3

# Forms
django-crispy-forms>=2.1
crispy-tailwind>=1.0

# File handling
Pillow>=10.2

# Real-time (Django Channels)
channels>=4.0
channels-redis>=4.2
daphne>=4.1

# Exports & Utilities
openpyxl>=3.1
icalendar>=5.0

# Testing
pytest>=8.0
pytest-django>=4.8

# Dev utilities
django-extensions>=3.2
